# Vitest Addon: Fix error message logic in set up file